refactor: Refactored into libs
